requirements_ml_processing.txt
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
ansible-base==2.10.8
anyio==4.9.0
apache-libcloud==3.2.0
appdirs==1.4.4
apprise==1.9.3
argcomplete==1.8.1
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
Babel==2.8.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
bokeh==3.7.3
Brotli==1.0.9
cachetools==5.5.2
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.0.3
click-plugins==1.1.1
cloudpickle==3.1.1
colorama==0.4.4
contourpy==1.3.2
coolname==2.2.0
cryptography==45.0.2
cycler==0.11.0
Cython==3.1.1
dateparser==1.2.1
dbus-python==1.2.18
decorator==4.4.2
Deprecated==1.2.18
distlib==0.3.8
distro==1.7.0
dnspython==2.1.0
docker==7.1.0
docker-pycreds==0.4.0
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.15.4
fonttools==4.29.1
fs==2.4.12
fsspec==2025.3.2
gast==0.5.2
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
greenlet==3.2.2
griffe==1.7.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
humanize==4.12.3
hyperframe==6.1.0
idna==3.3
importlib_metadata==8.6.1
iniconfig==2.1.0
jeepney==0.7.1
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jiter==0.9.0
jmespath==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keyring==23.5.0
kiwisolver==1.3.2
langchain==0.3.25
langchain-core==0.3.60
langchain-ollama==0.3.3
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.3.42
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.5.1
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
narwhals==1.40.0
netaddr==0.8.0
netifaces==0.11.0
networkx==2.4
ntlm-auth==1.4.0
numpy==1.24.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
olefile==0.46
ollama==0.4.8
openai==1.79.0
opentelemetry-api==1.33.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.8.0
Pillow==9.0.1
pipenv==2024.0.1
platformdirs==4.2.2
pluggy==1.5.0
ply==3.11
prefect==3.4.1
prometheus_client==0.22.0
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
py==1.10.0
pybind11==2.13.6
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.11.0
pydantic==2.11.5
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pygccxml==2.2.1
Pygments==2.19.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.17
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyRFC3339==1.1
pyroomacoustics==0.8.4
pyrsistent==0.18.1
pyspnego==0.11.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
python-socks==2.7.1
pythran==0.10.0
pytz==2025.2
pywinrm==0.3.0
pyxdg==0.27
PyYAML==6.0.2
pyzmq==22.3.0
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
requests-ntlm==1.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.25.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
schema==0.6.7
scikit-learn==1.6.1
scipy==1.8.0
SecretStorage==3.3.1
sentry-sdk==2.29.1
setproctitle==1.3.6
shellingham==1.5.4
simplejson==3.17.6
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
thrift==0.16.0
tiktoken==0.9.0
toml==0.10.2
tomli==2.2.1
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
triton==3.3.0
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ufoLib2==0.13.1
ujson==5.10.0
unicodedata2==14.0.0
urllib3==2.4.0
uuid7==0.1.0
uv==0.7.5
uvicorn==0.34.2
virtualenv==20.26.3
virtualenv-clone==0.3.0
wadllib==1.3.6
wandb==0.19.11
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
xdg==5.0.0
xmltodict==0.12.0
xyzservices==2025.4.0
zipp==3.21.0
zstandard==0.23.0
